refactor: merge `NiPointX.h` into `NiPoint.h`
