+++ CommonLibF4/include/RE/Bethesda/Actor.h
@@ -14,7 +14,7 @@
 #include "RE/Bethesda/IMovementInterface.h"
 #include "RE/Bethesda/TESObjectREFRs.h"
 #include "RE/NetImmerse/NiFlags.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
