+++ CommonLibF4/include/RE/Bethesda/BGSPrimitive.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
