+++ CommonLibF4/include/RE/Bethesda/BGSTextureSet.h
@@ -10,8 +10,7 @@
 #include "RE/Bethesda/TESForms.h"
 #include "RE/NetImmerse/NiAlphaProperty.h"
 #include "RE/NetImmerse/NiColor.h"
-#include "RE/NetImmerse/NiPoint2.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
