+++ CommonLibF4/include/RE/Bethesda/BSAudioManager.h
@@ -5,7 +5,7 @@
 #include "RE/Bethesda/BSTHashMap.h"
 #include "RE/Bethesda/BSTList.h"
 #include "RE/NetImmerse/NiAVObject.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
