+++ CommonLibF4/include/RE/Bethesda/BSShaderMaterial.h
@@ -5,7 +5,7 @@
 #include "RE/Bethesda/BSTSmartPointer.h"
 #include "RE/Bethesda/BSTextureDB.h"
 #include "RE/NetImmerse/NiColor.h"
-#include "RE/NetImmerse/NiPoint2.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
