+++ CommonLibF4/include/RE/Bethesda/BSTInterpolator.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RE/NetImmerse/NiPoint2.h"
+#include "RE/NetImmerse/NiPoint.h"
 
 namespace RE
 {
