+++ CommonLibF4/include/RE/Bethesda/FormComponents.h
@@ -9,7 +9,7 @@
 #include "RE/Bethesda/BSTSmartPointer.h"
 #include "RE/Bethesda/BSTTuple.h"
 #include "RE/Bethesda/MemoryManager.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiRefObject.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
