+++ CommonLibF4/include/RE/Bethesda/Interface3D.h
@@ -5,7 +5,7 @@
 #include "RE/Bethesda/BSTArray.h"
 #include "RE/Bethesda/UIShaderFXInfo.h"
 #include "RE/NetImmerse/NiColor.h"
-#include "RE/NetImmerse/NiPoint2.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
