+++ CommonLibF4/include/RE/Bethesda/PlayerControls.h
@@ -8,8 +8,7 @@
 #include "RE/Bethesda/BSTEvent.h"
 #include "RE/Bethesda/BSTSingleton.h"
 #include "RE/Bethesda/IMovementInterface.h"
-#include "RE/NetImmerse/NiPoint2.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 
 namespace RE
 {
