+++ CommonLibF4/include/RE/Bethesda/TESBoundAnimObjects.h
@@ -10,7 +10,7 @@
 #include "RE/Bethesda/TESBoundObjects.h"
 #include "RE/Bethesda/TESCondition.h"
 #include "RE/NetImmerse/NiColor.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 
 namespace RE
 {
