+++ CommonLibF4/include/RE/Bethesda/TESPackages.h
@@ -5,7 +5,7 @@
 #include "RE/Bethesda/BSTList.h"
 #include "RE/Bethesda/TESCondition.h"
 #include "RE/Bethesda/TESForms.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiSmartPointer.h"
 
 namespace RE
