+++ CommonLibF4/include/RE/Bethesda/Workshop.h
@@ -3,7 +3,7 @@
 #include "RE/Bethesda/BSExtraData.h"
 #include "RE/Bethesda/BSTArray.h"
 #include "RE/Bethesda/MemoryManager.h"
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 
 namespace RE
 {
