+++ CommonLibF4/include/RE/Fallout.h
@@ -278,9 +278,7 @@
 #include "RE/NetImmerse/NiNode.h"
 #include "RE/NetImmerse/NiObject.h"
 #include "RE/NetImmerse/NiObjectNET.h"
-#include "RE/NetImmerse/NiPoint2.h"
-#include "RE/NetImmerse/NiPoint3.h"
-#include "RE/NetImmerse/NiPoint4.h"
+#include "RE/NetImmerse/NiPoint.h"
 #include "RE/NetImmerse/NiProperty.h"
 #include "RE/NetImmerse/NiQuaternion.h"
 #include "RE/NetImmerse/NiRTTI.h"
