+++ CommonLibF4/include/RE/NetImmerse/NiBound.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RE/NetImmerse/NiPoint3.h"
+#include "RE/NetImmerse/NiPoint.h"
 
 namespace RE
 {
