+++ CommonLibF4/include/RE/NetImmerse/NiMatrix3.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include "RE/NetImmerse/NiPoint3.h"
-#include "RE/NetImmerse/NiPoint4.h"
+#include "RE/NetImmerse/NiPoint.h"
 
 namespace RE
 {
